requirements.txt
altair
numpy
pandas
pydeck
streamlit
python-dotenv
aiohttp